Combine Auth data, Add Custom hook of selectSort
