+++ .github/pull-request-template.md
@@ -18,11 +18,11 @@
 -
 -
 
-## 스크린샷
-
-![이미지 설명](이미지url)
-
 ## 멘토에게
 
 -
 -
+
+## 스크린샷
+
+![이미지 설명](이미지url)
