+++ src/components/Items/NavBar/SelectSort/SelectSort.jsx
@@ -1,25 +1,9 @@
-import { useState } from "react";
 import style from "./SelectSort.module.scss";
+import useSelectSort from "../../../../hooks/useSelectSort";
 
 const SelectSort = ({ sortLoad }) => {
-  const [currentSort, setCurrentSort] = useState("최신순");
-  const [sortBtn, setSortBtn] = useState(false);
-
-  // 정렬 선택버튼 토글
-  const sortPopupOn = () => {
-    setSortBtn(!sortBtn);
-  };
-
-  const sortPopupOff = (e, orderBy) => {
-    setSortBtn(false);
-    setCurrentSort(e.target.innerText);
-    sortSelect(orderBy);
-  };
-
-  // 정렬 선택 시 렌더링
-  const sortSelect = (orderBy) => {
-    sortLoad(orderBy);
-  };
+  const [currentSort, sortBtn, sortPopupOn, sortPopupOff] =
+    useSelectSort(sortLoad);
 
   return (
     <div className={style.selectContainer}>
